pyproject.toml
[project]
name = "job_runner"
version = "0.3.0"
description = "AI-powered end-to-end job application pipeline"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [{name = "Pickle-Pixel"}]
keywords = ["job-search", "automation", "ai", "resume", "apply", "job-application"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "httpx>=0.24",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "pypdf",
    "reportlab",
    "openai>=1.40",
    # Job board scraping (PyPI: python-jobspy, imports as jobspy). Core so discover works after any normal install.
    # Do not install the unrelated ``jobspy`` package on PyPI — it breaks imports.
    # 1.1.82 wheel was published incomplete (missing indeed/, bayt/, etc.); exclude until fixed upstream.
    "python-jobspy>=1.1.0,!=1.1.82",
    ]

[project.optional-dependencies]
webui = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.0",
]
# Kept for backward compatibility; python-jobspy is already a core dependency.
discover = ["python-jobspy>=1.1.0,!=1.1.82"]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
job_runner = "job_runner.cli:app"

[project.urls]
Homepage = "https://github.com/Pickle-Pixel/job_runner"
Repository = "https://github.com/Pickle-Pixel/job_runner"
Issues = "https://github.com/Pickle-Pixel/job_runner/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_runner"]

[tool.hatch.build]
artifacts = ["src/job_runner/config/*.yaml"]

[tool.ruff]
target-version = "py311"
line-length = 120
